separação do controller Categorias
